docs/#GCP 講義/examples/py-helloworld-storage/requirements.txt
Flask==3.0.0
gunicorn==22.0.0
Werkzeug==3.0.1
cloudevents==1.10.1
google-cloud-storage==2.16.0
